Add software character generator print function.
So far, the LED sign library has contained only pieces of a
software character generator print function. The function itself
was in the daemon. This puts a print function in the library.

The function is complex because it handles carriage returns,
line feeds and looping. Unlike the hardware character generator,
it does not first clear the screen. Notification and exit polling
is still incomplete.

The daemon still uses its own code, which is tightly integrated
with the way the message is stored in chains of blocks. That can
be eliminated by replacing the zero terminated string option with
an option for specifying data length as a switch for commands.