Use serial port output flush to decrease VU display latency.
Past experiments have shown a lot of serial output buffering in
Linux. This adds serial a flush to avoid queueing up VU data.

The flush is before VU data is updated because that should
minimize performance impact. At that point, with repeated vu
updates the maximum length of time has passed since the last
update, so the wait should be shortest.