Create platform-independent RGB lamp visualization code.
This adds platform-independent RGB lamp visualization code and
uses it for the Audacious plugin. The spectrum is split in 3
parts which are set so that when pink noise is played, the parts
all sum to approximately the same value.

The Winamp plugin has not been updated or tested, and its build
process may be broken.

This removes the separate aud_rgb code because it is not needed.
Also, it was broken because it assumed Audacious returned 512
frequency bins, and Audacious returns only 256.