Include font in LED sign daemon, so file is not needed at runtime
In Windows the font was loaded from a resource, but elsewhere the
font file needed to be in the current directory. This led to
confusing error messages about being unable to initialize hardware.
Now font is transformed to C via font2c and added via #include.
LED sign API was changed to make font parameters pointer to const
where appropriate.