Exprimental automatic gain control, enabled in Windows standalone
I'm not convinced that AGC is a good thing in general as a clear
link between loudness and brightness feels right. So I'm mainly
adding this because Stereo Mix in Windows records sound scaled
down by its recording volume and by playback volume. AGC is better
than having dim light because of that. It's also available for
experimenting in other builds.