Change timeout API to ask for length, and use for lamp power
Previously, timeout API asked hardware for yes/no about whether
timeout calls are needed. Currently it asks for number of seconds
or 0 if not needed. This allows timeouts to be used for turning
off lamp power supply. Also, it gives timeout length decision to
serio, where it makes most sense.