Shorten HASS name and ID, config via /usr/local/etc/coloranimd.conf
Home Assistant 2023.11.1 concatenates MQTT name to device name,
except when the names are the same. So, make the names the same.

At build time /usr/local/etc/coloranimd.conf can be used to
configure coloranimd. There seems to be no need to complicate
things by including runtime configuration for this.