Make RGB light drivers into shared libraries loaded by rgbvis
This avoids the need for multiple rgbvis binaries, and allows
multiple lights to be controlled at the same time. The new
rgbm_multi.c handles loading and passing of data to light drivers.
The 1st command line argument to rgbvis can specify multiple
drivers separated by a + sign, like lamp+strip. This is not tested
in Windows.